Add a folder per language and a folder per stage
Currently, the only language folder is Python, but the idea is
to have one folder per language used to implement the min-hexagonal.

Although I'm keeping the narrative for the changes between stages in
the git log, the final product should include it, and it is now
represented in 3 folder, one per stage.